moved ParamCache to another file
