+++ price-feeder/oracle/param_test.go
@@ -1,0 +1,65 @@
+package oracle
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	oracletypes "github.com/umee-network/umee/v2/x/oracle/types"
+)
+
+func TestParamCacheIsParamsOutdated(t *testing.T) {
+	testCases := map[string]struct {
+		paramCache        ParamCache
+		currentBlockHeigh int64
+		expected          bool
+	}{
+		"Params Nil": {
+			paramCache: ParamCache{
+				params:           nil,
+				lastUpdatedBlock: 0,
+			},
+			currentBlockHeigh: 10,
+			expected:          true,
+		},
+		"currentBlockHeigh < cacheOnChainBlockQuantity": {
+			paramCache: ParamCache{
+				params:           &oracletypes.Params{},
+				lastUpdatedBlock: 0,
+			},
+			currentBlockHeigh: 199,
+			expected:          false,
+		},
+		"currentBlockHeigh < lastUpdatedBlock": {
+			paramCache: ParamCache{
+				params:           &oracletypes.Params{},
+				lastUpdatedBlock: 205,
+			},
+			currentBlockHeigh: 203,
+			expected:          true,
+		},
+		"Outdated": {
+			paramCache: ParamCache{
+				params:           &oracletypes.Params{},
+				lastUpdatedBlock: 200,
+			},
+			currentBlockHeigh: 401,
+			expected:          true,
+		},
+		"Limit to keep in cache": {
+			paramCache: ParamCache{
+				params:           &oracletypes.Params{},
+				lastUpdatedBlock: 200,
+			},
+			currentBlockHeigh: 400,
+			expected:          false,
+		},
+	}
+
+	for name, tc := range testCases {
+		tc := tc
+
+		t.Run(name, func(t *testing.T) {
+			require.Equal(t, tc.expected, tc.paramCache.IsParamsOutdated(tc.currentBlockHeigh))
+		})
+	}
+}
